chore: refactor of files path for better organization
