+++ src/authz-module/components/RenderAdminRole.tsx
@@ -1,5 +1,5 @@
 import { useIntl } from '@edx/frontend-platform/i18n';
-import messages from './messages';
+import messages from '../audit-user/messages';
 
 interface RenderAdminRoleProps {
   role: string;
