+++ src/authz-module/components/RenderPermissionColumn.tsx
@@ -1,6 +1,6 @@
 import { Icon } from '@openedx/paragon';
 import { RolePermission } from 'types';
-import ResourceTooltip from '../components/ResourceTooltip';
+import ResourceTooltip from './ResourceTooltip';
 
 interface ExtendedRolePermission extends RolePermission {
   icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
