+++ src/authz-module/hooks/useQuerySettings.test.tsx
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react';
-import { QuerySettings } from '../data/api';
+import { QuerySettings } from '@src/authz-module/data/api';
 import { useQuerySettings } from './useQuerySettings';
 
 describe('useQuerySettings', () => {
